Refactor modules and remove deprecated macro_generator module
