--- anypytools/utils/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
-from builtins import *
-
