+++ setup.py
@@ -12,13 +12,15 @@ setup(
     name='AnyPyTools',
     version='0.9.7',
     install_requires=['future', 'numpy'],
-    py_modules=['anypytools.abcutils', 'anypytools.h5py_wrapper',
-                'anypytools.datautils', 'anypytools.genereate_macros',
-                'anypytools.pytest_plugin',
-                'anypytools.tools',
-                'anypytools.utils.blaze_converter',
-                'anypytools.pygments_plugin.anyscript_lexer',
-                'anypytools.pygments_plugin.anyscript_style'],
+    py_modules=[
+        'anypytools.abcutils',
+        'anypytools.h5py_wrapper',
+        'anypytools.datautils',
+        'anypytools.pytest_plugin',
+        'anypytools.tools',
+        'anypytools.blaze_converter',
+        'anypytools.pygments_plugin.anyscript_lexer',
+        'anypytools.pygments_plugin.anyscript_style'],
     packages=['anypytools'],
     package_data={'anypytools': ['test_models/Demo.Arm2D.any', 'pygments_plugin/*.txt']},
     # the following makes a plugin available to pytest
